listening-comp/backend/requirements.txt
faiss-cpu
streamlit
boto3
youtube-transcript-api
googletrans==4.0.0-rc1
httpx==0.13.3